pyproject.toml
[tool.poetry]
name = "parsertang"
version = "0.1.0"
description = "Realtime CEX stablecoin arbitrage parser with WS-first streaming, liquidity filter, fees-aware net profit and Telegram alerts"
authors = ["You <you@example.com>"]
readme = "README.md"
package-mode = true
packages = [{ include = "parsertang", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
ccxt = "^4.3.17"
# ccxtpro is commercial; include optionally if available
ccxtpro = "^1.0.1"
pandas = "^2.2.2"
python-telegram-bot = "^21.2"
tenacity = "^8.2.3"
pydantic = "^2.8.0"
pydantic-settings = "^2.4.0"
protobuf = "5.29.5"
aiohttp-socks = "^0.10.0"  # SOCKS5 proxy support for WebSocket (HTX geo-block bypass)
pysocks = "^1.7.1"  # SOCKS5 proxy support for sync ccxt REST API

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-asyncio = "^1.2.0"
ruff = "^0.8.0"
pyright = "^1.1.390"
pre-commit = "^4.5.1"

[tool.poetry.group.diagnostic.dependencies]
typer = "^0.9.0"
rich = "^13.0.0"
tqdm = "^4.66.0"
pytest-mock = "^3.12.0"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
